+++ src/app/storage-handler.ts
@@ -1,0 +1,4 @@
+export interface StorageHandler<K, V> {
+    save(id: K, data: V): void;
+    load(id: K): V;
+}